% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulver.R
\name{read_databel}
\alias{read_databel}
\title{Reads a \code{databel} object file in table format and returns a matrix of type "double"}
\usage{
read_databel(file)
}
\arguments{
\item{file}{file name of a \code{databel} object (ending: ".fvi") or text file}
}
\value{
matrix of type "double"
}
\description{
Reads a \code{databel} object file in table format and returns a matrix of type "double"
}
\references{
\url{http://www.genabel.org/packages/DatABEL}
}
