% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulver.R
\name{write_databel}
\alias{write_databel}
\title{Converts matrix of type "double" to a \code{databel} object file}
\usage{
write_databel(data, file)
}
\arguments{
\item{data}{matrix of type "double"}

\item{file}{file name of a \code{databel} object (extension: ".fvi") or text file}
}
\description{
Converts matrix of type "double" to a \code{databel} object file
}
\references{
\url{www.genabel.org/packages/DatABEL}
}
